requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
ytmusicapi==1.11.5
yt-dlp>=2025.01.15
curl-cffi==0.11.0
httpx==0.28.1
python-dotenv
pydantic-settings
# Database
sqlalchemy==2.0.23
# psycopg2-binary removed for SQLite
email-validator==2.2.0
alembic==1.13.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
anyio
diskcache
